routes get protected and project finished

+++ mycontacts-backend/controllers/contactController.js
@@ -3,7 +3,7 @@ const Contact = require('./../models/contactModel')
 
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({ user_id: req.user.id })
     res.status(200).json(contacts);
 })
 
@@ -17,7 +17,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     })
     res.status(201).json(contact);
 })
@@ -39,6 +40,11 @@ const updateContact = asyncHandler(async (req, res) => {
         throw new Error("contact not found")
     }
 
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User dont have permission to update other user contacts")
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -54,6 +60,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("contact not found")
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User dont have permission to delete other user contacts")
     }
 
     await Contact.findByIdAndDelete(req.params.id);
